\name{profitCubaSersic}
\alias{profitCubaSersic}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Exact 2D Sersic Model Image Creation
}
\description{
This function computes the exact pixel integrals for a given model image. This is very slow compared to the low-level \code{\link{profitMakeSersic}} or high-level \code{\link{profitMakeModel}}, but it is useful for checking model creation tuning (i.e. the degree to which speed can be increased without overly harming accuracy). Tests with this function were used to tune \code{\link{profitMakeModel}}.
}
\usage{
profitCubaSersic(xcen = 0, ycen = 0, mag = 15, re = 1, nser = 1, ang = 0, axrat = 1,
box = 0, xlim = c(-10, 10), ylim = c(-10, 10), dim = c(20, 20), acc = 1e-3)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xcen}{
x centre of the 2D Sersic profile (can be fractional pixel positions).
}
  \item{ycen}{
y centre of the 2D Sersic profile (can be fractional pixel positions).
}
  \item{mag}{
Total magnitude of the 2D Sersic profile. Converted to flux using flux=10^(-0.4*(mag-magzero)).
}
  \item{re}{
Effective radius of the Sersic profile.
}
  \item{nser}{
Sersic index of the Sersic profile.
}
  \item{ang}{
The orientation of the major axis of the Sersic profile in degrees. When plotted as an R image the angle (theta) has the convention that 0= | (vertical), 45= \, 90= - (horizontal), 135= /, 180= | (vertical). Values outside the range 0 <= ang <= 180 are allowed, but these get recomputed as ang = ang.
}
  \item{axrat}{
Axial ratio of the Sersic profile defined as minor-axis/major-axis, i.e. 1 is a circle and 0 is a line.
}
  \item{box}{
The boxiness of the Sersic profile that traces contours of iso-flux, defined such that r[mod]=(x^(2+box)+y^(2+box))^(1/(2+box)). When box=0 the iso-flux contours will be normal ellipses, but modifications between -1<box<1 will produce visually boxy distortions. Negative values have a pin-cushion effect, whereas positive values have a barrel effect (the major and minor axes staying fixed in all cases).
}
  \item{xlim}{
The x pixel range of the image to generate the image on. Typically this should be c(0,Nx).
}
  \item{ylim}{
The y pixel range of the image to generate the image on. Typically this should be c(0,Ny).
}
  \item{dim}{
The dimensions of the image to be generated. Typically this should be c(Nx,Ny).
}
  \item{acc}{
The target relative accuracy of the cubature integral. The relative flux in a pixel will be no more inaccurate than this specified figure. Making this smaller will increase (the already long) computation time. The default of \option{acc}=1e-3 means computed fluxes are accurate to at least 0.1\% (and usually much better than this figure).
}
}
\details{
This function uses the Cuba package to make an accurate (but expensive) cubature integral. This function is only used to test the accuracy of ProFit models generated by \code{\link{profitMakeSersic}} and \code{\link{profitMakeModel}}.
}
\value{
Matrix; contains the flux values of the specified model image. Dimensions dim.
}
\author{
Aaron Robotham
}

\seealso{
\code{\link{profitMakeSersic}}, \code{\link{profitMakeModel}}
}
\examples{
model = list(
	sersic = list(
		xcen = 10,
		ycen = 10,
		mag = 15,
		re = 2,
		nser = 4,
		ang = 30,
		axrat = 0.5,
		box = 0
	)
)

xlim=c(0,20)
ylim=c(0,20)
dim=c(20,20)

tempExact=profitCubaSersic(xcen=model$sersic$xcen, ycen=model$sersic$ycen,
mag=model$sersic$mag, re=model$sersic$re, nser=model$sersic$nser, ang=model$sersic$ang,
axrat=model$sersic$axrat, box=model$sersic$box, xlim=xlim, ylim=ylim, dim=dim)
tempProFit=profitMakeModel(model, dim=dim)$z

#The absolute differences between the exact and approximate ProFit model image:
magimage(tempExact-tempProFit)
#The relative differences between the exact and approximate ProFit model image:
magimage((tempExact-tempProFit)/tempExact)
#They differ by no more the 1\% in flux for any pixel, and in general much less than that:
hist((tempExact-tempProFit)/tempExact)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
